fix: address PR review feedback
- Add header redaction to EventTracker debug logger (Api-Key, Merchant-Id)
- Add PAYMENT_CANCELLED error code for analytics distinction from PAYMENT_FAILED
- Change fiatAmount from Int to Long to prevent silent truncation for large amounts
- Replace hardcoded colors in PaymentSuccessScreen with WCTheme tokens

+++ product/pos/src/main/kotlin/com/walletconnect/pos/Pos.kt
@@ -72,7 +72,7 @@ object Pos {
         val referenceId: String?,
         val status: TransactionStatus,
         val txHash: String?,
-        val fiatAmount: Int?,
+        val fiatAmount: Long?,
         val fiatCurrency: String?,
         val tokenAmount: String?,
         val tokenSymbol: String?,
